% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_of_sep.R
\name{classify_sep_dest}
\alias{classify_sep_dest}
\title{Classify modes of sep using extracted groups & destination}
\usage{
classify_sep_dest(df, prefix = mode_sep, top_suffix = first,
                         dests = destination)
}
\description{
Classify modes of sep using extracted groups & destination
}
\keyword{internal}
