% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{create_journey_table}
\alias{create_journey_table}
\title{Extracts the unique journey_id and assigns an order for each of the patient's
journeys and calculates the time between subsequent journeys}
\usage{
create_journey_table(
  journeys,
  last_journey_number = NULL,
  ...,
  .default_dots = rlang::exprs(journey_key = JOURNEY_KEY, journey_start =
    JOURNEY_START_DT_TM, journey_end = JOURNEY_END_DT_TM, journey_day = JOURNEY_DAYS,
    updt_dttm = UPDT_DT_TM, encntr_key_end = ENCNTR_ID_END, person_key = PERSON_ID,
    last_journey_order = LASt_JOURNEY_ORDER)
)
}
\arguments{
\item{journeys}{the journey df}

\item{last_journey_number}{a df containing journey_key and the most recent journey rank number}

\item{...}{see section Column Specification}
}
\value{
the df,

updt_dt_tm the most recent update of the encounters within the journey
}
\description{
Extracts the unique journey_id and assigns an order for each of the patient's
journeys and calculates the time between subsequent journeys
}
\seealso{
Other journey_building: 
\code{\link{journey_length}()},
\code{\link{journey_order}()},
\code{\link{link_encounters_by_timestamps}()},
\code{\link{link_encs_in_groups}()},
\code{\link{relink_apac_to_journeys}()},
\code{\link{terminate_journeys_by}()}
}
\concept{journey_building}
\keyword{internal}
