% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{create_visit_journey}
\alias{create_visit_journey}
\title{Extract first and last visit for each patient}
\usage{
create_visit_journey(visit_form)
}
\arguments{
\item{visit_form}{a df containing visit forms, each patient can have multiple visit forms}
}
\value{
a df, where each row represents a patient and information about their first and last visit
}
\description{
Extracts the dates, NYHA classification and the type of visit for each patients
}
\seealso{
\code{\link[=process_hfreferral_form]{process_hfreferral_form()}} for more details about \code{visit_form}

Other hfhomevisit: 
\code{\link{create_visit_trajectory}()},
\code{\link{extract_hf_symptoms}()},
\code{\link{get_homevisit_forms}()},
\code{\link{process_homevisit_form}()}
}
\concept{hfhomevisit}
