% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{create_visit_trajectory}
\alias{create_visit_trajectory}
\title{Process the MACARF_VISIT_TYPE field}
\usage{
create_visit_trajectory(hfhomevisit)
}
\arguments{
\item{hfhomevisit}{a dataframe where each row represents a form and the column is each field}
}
\value{
a dataframe where each row represents patient level information
}
\description{
Extracts the visit type of each form
and assigns the information to a column for each patient.
Then assigns a class to each patient, which describes which part of their
trajectory in the program
e.g. if \code{HAS_HOME_VISIT = TRUE}, \code{HAS_PHONE_CALL_1_12_VISIT = FALSE},
\code{HAS_PHONE_CALL_3_12_VISIT = FALSE}, \code{HAS_PHONE_CALL_6_12_VISIT =TRUE}
the patient would be classified as being in "Phone Call 6/12" of their journey
}
\details{
Calculates the number of home visit each person has
}
\seealso{
\code{\link[=process_homevisit_form]{process_homevisit_form()}} for details about how this field is processed

Other hfhomevisit: 
\code{\link{create_visit_journey}()},
\code{\link{extract_hf_symptoms}()},
\code{\link{get_homevisit_forms}()},
\code{\link{process_homevisit_form}()}
}
\concept{hfhomevisit}
