% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_query}
\alias{execute_query}
\title{Execute the query}
\usage{
execute_query(query)
}
\arguments{
\item{query}{a string, query to execute}
}
\description{
Execute the query
}
\seealso{
Other data-utility: 
\code{\link{get_first_non_na}()},
\code{\link{load_data}()},
\code{\link{load_db}()},
\code{\link{str_to_datetime}()}
}
\concept{data-utility}
