% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_echos}
\alias{extract_echos}
\title{extracts the most recent echo order to the encounter}
\usage{
extract_echos(df, df_dttm = "DISCH_DT_TM")
}
\description{
extracts the most recent echo order to the encounter
}
