% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_form_result}
\alias{extract_form_result}
\title{extracts the result and bin pathology result}
\usage{
extract_form_result(
  df,
  regex,
  prefix,
  breaks = NULL,
  labels = NULL,
  is_numeric = FALSE
)
}
\description{
extracts the result and bin pathology result
}
\seealso{
\code{\link[=process_inpatient]{process_inpatient()}} for df

Other hfinpatient: 
\code{\link{extract_age_los}()},
\code{\link{extract_alert}()},
\code{\link{extract_diagnosis}()},
\code{\link{extract_discharge_meds}()},
\code{\link{extract_echos}()},
\code{\link{extract_hfenrolment_form}()},
\code{\link{extract_hfreferral_form}()},
\code{\link{extract_lvef_result}()},
\code{\link{extract_path_result}()},
\code{\link{extract_patho_result}()},
\code{\link{get_most_recent_order}()},
\code{\link{process_inpatient}()}
}
\concept{hfinpatient}
