% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{extract_hf_symptoms}
\alias{extract_hf_symptoms}
\title{Process HEART_FAILURE_SYMPTOMS field in Homevisit form}
\usage{
extract_hf_symptoms(df, column = "HEART_FAILURE_SYMPTOMS")
}
\arguments{
\item{df}{a dataframe where each row represents a homevisit form and each column the fields}

\item{column}{the column name containing the \code{HEART_FAILURE_SYMPTOMS}}
}
\description{
Exytracts the HEART_FAILURE_SYMPTOMS column
and separates the value in this field into specific, separate columns
}
\seealso{
\code{\link[=process_homevisit_form]{process_homevisit_form()}} for more details about \code{visit_form}

Other hfhomevisit: 
\code{\link{create_visit_journey}()},
\code{\link{create_visit_trajectory}()},
\code{\link{get_homevisit_forms}()},
\code{\link{process_homevisit_form}()}
}
\concept{hfhomevisit}
