% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_lvef_result}
\alias{extract_lvef_result}
\title{extracts the LVEF result}
\usage{
extract_lvef_result(df, col = "LV_SUMMARY", new_col = "LVEF_STRIP")
}
\arguments{
\item{df}{a dataframe containing the LV summary}

\item{col}{the column that contains the LV summary}

\item{new_col}{the name of the newly created column to store}
}
\description{
extracts the LVEF result
}
\seealso{
\code{\link[=get_lvef_result_to_strip]{get_lvef_result_to_strip()}} for df

Other hfinpatient: 
\code{\link{extract_age_los}()},
\code{\link{extract_alert}()},
\code{\link{extract_diagnosis}()},
\code{\link{extract_discharge_meds}()},
\code{\link{extract_echos}()},
\code{\link{extract_form_result}()},
\code{\link{extract_hfenrolment_form}()},
\code{\link{extract_hfreferral_form}()},
\code{\link{extract_path_result}()},
\code{\link{extract_patho_result}()},
\code{\link{get_most_recent_order}()},
\code{\link{process_inpatient}()}
}
\concept{hfinpatient}
