% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_path_result}
\alias{extract_path_result}
\title{extracts the result and bin pathology result for a given rest}
\usage{
extract_path_result(df, regex, prefix, breaks = NULL, labels = NULL)
}
\description{
extracts the result and bin pathology result for a given rest
}
