% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_patho_result}
\alias{extract_patho_result}
\title{Process HF related pathology}
\usage{
extract_patho_result(
  df,
  use_default_person_list = TRUE,
  df_dttm = "DISCH_DT_TM"
)
}
\arguments{
\item{df}{a dataframe containing encounters}

\item{use_default_person_list}{a boolean
determines whether the list of ids is based on the given
df or based on the default query}

\item{df_dttm, }{the column in df to coompare the timestamp of orders against}
}
\description{
Processes HF related pathology for a list of patients
}
\seealso{
\code{\link[=process_inpatient]{process_inpatient()}} for df

Other hfinpatient: 
\code{\link{extract_age_los}()},
\code{\link{extract_alert}()},
\code{\link{extract_diagnosis}()},
\code{\link{extract_discharge_meds}()},
\code{\link{extract_echos}()},
\code{\link{extract_form_result}()},
\code{\link{extract_hfenrolment_form}()},
\code{\link{extract_hfreferral_form}()},
\code{\link{extract_lvef_result}()},
\code{\link{extract_path_result}()},
\code{\link{get_most_recent_order}()},
\code{\link{process_inpatient}()}
}
\concept{hfinpatient}
