% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmanagement.R
\name{get_aetiology}
\alias{get_aetiology}
\title{Processes Aetiology information in enrolment form}
\usage{
get_aetiology(df)
}
\arguments{
\item{df}{a dataframe containing enrolment forms}
}
\description{
Converts the aetiology fields in the enrolment into separate categories
}
\seealso{
\code{\link[=process_hfmanagement_form]{process_hfmanagement_form()}} for df,
\code{\link[=get_precipitants]{get_precipitants()}} for processing precipitant field

Other hfmanagement: 
\code{\link{get_ecg_data}()},
\code{\link{get_ecg_on_admit}()},
\code{\link{get_precipitants}()},
\code{\link{process_hfmanagement_form}()}
}
\concept{hfmanagement}
