% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_age_bin}
\alias{get_age_bin}
\title{Define age categories}
\usage{
get_age_bin()
}
\description{
Define age categories
}
\seealso{
Other definition: 
\code{\link{get_bnp_bin}()},
\code{\link{get_ferritin_bin}()},
\code{\link{get_hfenrolment_form_pattern}()},
\code{\link{get_hfhomevisit_form_pattern}()},
\code{\link{get_lvef_bin}()}
}
\concept{definition}
