% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_bnp_bin}
\alias{get_bnp_bin}
\title{Define bnp categories}
\usage{
get_bnp_bin()
}
\description{
Define bnp categories
}
\seealso{
Other definition: 
\code{\link{get_age_bin}()},
\code{\link{get_ferritin_bin}()},
\code{\link{get_hfenrolment_form_pattern}()},
\code{\link{get_hfhomevisit_form_pattern}()},
\code{\link{get_lvef_bin}()}
}
\concept{definition}
