% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Get the connection to the database}
\usage{
get_db_connection()
}
\description{
Get the connection to the database
}
\seealso{
Other data-utility: 
\code{\link{execute_query}()},
\code{\link{execute_write_to_db}()},
\code{\link{get_first_non_na}()},
\code{\link{load_data}()},
\code{\link{load_db}()},
\code{\link{str_to_datetime}()}
}
\concept{data-utility}
