% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_digit_regex}
\alias{get_digit_regex}
\title{Define the regex for extracting digits}
\usage{
get_digit_regex()
}
\description{
Define the regex for extracting digits
}
\seealso{
Other regex: 
\code{\link{get_lvef_regex}()}
}
\concept{regex}
