% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_discharge_referral}
\alias{get_discharge_referral}
\title{get the most recently updated note for each encounter}
\usage{
get_discharge_referral()
}
\description{
get the most recently updated note for each encounter
}
\seealso{
Other query: 
\code{\link{get_acd_problem_query}()},
\code{\link{get_disch_med_query}()},
\code{\link{get_encounter_journey_query}()},
\code{\link{get_hf_alert_query}()},
\code{\link{get_hfenrolment_form_query}()},
\code{\link{get_hfreferral_form_query}()},
\code{\link{get_historical_hf_diag_query}()},
\code{\link{get_previous_encounter_query}()},
\code{\link{get_referred_to_facility_query}()}
}
\concept{query}
