% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmanagement.R
\name{get_ecg_data}
\alias{get_ecg_data}
\title{Processes LVEF information in enrolment form}
\usage{
get_ecg_data(df)
}
\arguments{
\item{df}{a dataframe containing enrolment forms}
}
\description{
Converts the LVEF fields in the enrolment form into separate bins
and extracts the ECG date
}
\seealso{
\code{\link[=process_hfmanagement_form]{process_hfmanagement_form()}} for df

Other hfmanagement: 
\code{\link{get_aetiology}()},
\code{\link{get_ecg_on_admit}()},
\code{\link{get_precipitants}()},
\code{\link{process_hfmanagement_form}()}
}
\concept{hfmanagement}
