% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmanagement.R
\name{get_ecg_on_admit}
\alias{get_ecg_on_admit}
\title{Processes ECG information in enrolment form}
\usage{
get_ecg_on_admit(df)
}
\arguments{
\item{df}{a dataframe containing enrolment forms}
}
\description{
Converts the ECG fields in the enrolment into separate categories
}
\seealso{
\code{\link[=process_hfmanagement_form]{process_hfmanagement_form()}} for df,
\code{\link[=get_precipitants]{get_precipitants()}} for processing precipitant field

Other hfmanagement: 
\code{\link{get_aetiology}()},
\code{\link{get_ecg_data}()},
\code{\link{get_precipitants}()},
\code{\link{process_hfmanagement_form}()}
}
\concept{hfmanagement}
