% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_encounter_journey_query}
\alias{get_encounter_journey_query}
\title{the query to extract the encounters that haven't been allocated a journey id}
\usage{
get_encounter_journey_query()
}
\description{
`join_tab` extracts encounters from the encounter table and encounter history table
`last_journey` gets the last journey currently for each patient
`last_encounter` gets the encounters associated for those journeys
using the above, we extract the encounters that haven't been allocated a journey key
and the most recent encounters for each patient, just in case their journeys
are continued with the most recent set of new encounters
}
\details{
be aware that we assume that each encounter can only be associated with one person_id
across the encounter and encounter history tables
}
