% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_ferritin_bin}
\alias{get_ferritin_bin}
\title{Define ferritin categories}
\usage{
get_ferritin_bin()
}
\description{
Define ferritin categories
}
\seealso{
Other definition: 
\code{\link{get_age_bin}()},
\code{\link{get_bnp_bin}()},
\code{\link{get_hfenrolment_form_pattern}()},
\code{\link{get_hfhomevisit_form_pattern}()},
\code{\link{get_lvef_bin}()}
}
\concept{definition}
