% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_non_na}
\alias{get_first_non_na}
\title{Get first element excluding na}
\usage{
get_first_non_na(x)
}
\description{
Get first element excluding na
}
\seealso{
Other data-utility: 
\code{\link{execute_query}()},
\code{\link{execute_write_to_db}()},
\code{\link{get_db_connection}()},
\code{\link{load_data}()},
\code{\link{load_db}()},
\code{\link{str_to_datetime}()}
}
\concept{data-utility}
