% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_hf_encounter_query}
\alias{get_hf_encounter_query}
\title{Queries Encounters with HF Diagnosis}
\usage{
get_hf_encounter_query()
}
\description{
Query to extract encounters with any HF diagnosis (either primary or additional)
based on the flag qualifier \code{Q4}
}
\seealso{
Other query: 
\code{\link{get_acd_problem_query}()},
\code{\link{get_diagnosis_query}()},
\code{\link{get_disch_med_query}()},
\code{\link{get_discharge_referral}()},
\code{\link{get_echos_query}()},
\code{\link{get_encounter_journey_query}()},
\code{\link{get_hf_alert_query}()},
\code{\link{get_hfenrolment_form_query}()},
\code{\link{get_hfhomevisit_form_query}()},
\code{\link{get_hfreferral_form_query}()},
\code{\link{get_historical_hf_diag_query}()},
\code{\link{get_pathology_query}()},
\code{\link{get_previous_encounter_query}()},
\code{\link{get_referred_to_facility_query}()}
}
\concept{query}
