% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_hfhomevisit_form_pattern}
\alias{get_hfhomevisit_form_pattern}
\title{Define homevisit form string}
\usage{
get_hfhomevisit_form_pattern()
}
\description{
Define homevisit form string
}
\seealso{
Other definition: 
\code{\link{get_age_bin}()},
\code{\link{get_bnp_bin}()},
\code{\link{get_ferritin_bin}()},
\code{\link{get_hfenrolment_form_pattern}()},
\code{\link{get_lvef_bin}()}
}
\concept{definition}
