% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_hfhomevisit_form_query}
\alias{get_hfhomevisit_form_query}
\title{Queries HF Homevisit form}
\usage{
get_hfhomevisit_form_query()
}
\description{
Queries the HF homevisit form and all fields in that form
The form title exists in \code{DCP_FORMS_ACTIVITY}
and the fields of the form exists in \code{FORMS_EVENT}
joined by \code{PARENT_ENTITY_ID} \code{PARENT_EVENT_ID} respectively
}
\seealso{
Other query: 
\code{\link{get_acd_problem_query}()},
\code{\link{get_diagnosis_query}()},
\code{\link{get_disch_med_query}()},
\code{\link{get_discharge_referral}()},
\code{\link{get_echos_query}()},
\code{\link{get_encounter_journey_query}()},
\code{\link{get_hf_alert_query}()},
\code{\link{get_hf_encounter_query}()},
\code{\link{get_hfenrolment_form_query}()},
\code{\link{get_hfreferral_form_query}()},
\code{\link{get_historical_hf_diag_query}()},
\code{\link{get_pathology_query}()},
\code{\link{get_previous_encounter_query}()},
\code{\link{get_referred_to_facility_query}()}
}
\concept{query}
