% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{get_homevisit_forms}
\alias{get_homevisit_forms}
\title{A helper function to extract homevisit forms}
\usage{
get_homevisit_forms(forms = NULL, dcp_forms_activity = NULL)
}
\description{
Determines whether a query or the function will be used to extract the homevisit forms
}
\seealso{
Other hfhomevisit: 
\code{\link{create_visit_journey}()},
\code{\link{create_visit_trajectory}()},
\code{\link{extract_hf_symptoms}()},
\code{\link{process_homevisit_form}()}
}
\concept{hfhomevisit}
