% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_lvef_bin}
\alias{get_lvef_bin}
\title{Define lvef categories}
\usage{
get_lvef_bin()
}
\description{
Define lvef categories
}
\seealso{
Other definition: 
\code{\link{get_age_bin}()},
\code{\link{get_bnp_bin}()},
\code{\link{get_ferritin_bin}()},
\code{\link{get_hfenrolment_form_pattern}()},
\code{\link{get_hfhomevisit_form_pattern}()}
}
\concept{definition}
