% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_definitions.R
\name{get_lvef_regex}
\alias{get_lvef_regex}
\title{Define the regex for LVEF}
\usage{
get_lvef_regex()
}
\description{
Define the regex for LVEF
}
\seealso{
Other regex: 
\code{\link{get_digit_regex}()}
}
\concept{regex}
