% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfenrolment.R
\name{get_n_precipitants}
\alias{get_n_precipitants}
\title{Computes the number of precipiants and possible precipitants}
\usage{
get_n_precipitants(
  hfmanagement,
  HFPRECIPITANTS = "HFPRECIPITAMTS",
  HFPOSSIBLEPRECIPITANTS = "HFPOSSIBLEPRECIPITANTS",
  PERSON_ID = "PERSON_ID"
)
}
\arguments{
\item{hfmanagement}{a df containing form fields of the enrolment form}
}
\value{
a df, the hfmanagemet table with the number of precipitants and possible precipitants
}
\description{
Computes the number of precipiants and possible precipitants in the hfmanagement form
Field is considered missing when both possible and precipiants is left empty
Number of precipiants is considered 0 when no known precipiants is field
}
\seealso{
\code{\link[=process_hfenrolment_form]{process_hfenrolment_form()}}

Other hfenrolment: 
\code{\link{get_enrolment_fields}()},
\code{\link{process_hfenrolment_form}()},
\code{\link{process_hfmanagement}()},
\code{\link{summarise_enrolment}()}
}
\concept{hfenrolment}
