% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_pathology_query}
\alias{get_pathology_query}
\title{Queries BNP orders}
\usage{
get_pathology_query(person_id = NULL)
}
\arguments{
\item{person_id}{a list of person_id's
if \code{NULL} will query encounters with HF diagnosis}
}
\description{
Returns a query for BNP orders
either for HF encounters
or for the list of \code{person_id}
}
\seealso{
Other query: 
\code{\link{get_acd_problem_query}()},
\code{\link{get_diagnosis_query}()},
\code{\link{get_disch_med_query}()},
\code{\link{get_discharge_referral}()},
\code{\link{get_echos_query}()},
\code{\link{get_encounter_journey_query}()},
\code{\link{get_hf_alert_query}()},
\code{\link{get_hf_encounter_query}()},
\code{\link{get_hfenrolment_form_query}()},
\code{\link{get_hfhomevisit_form_query}()},
\code{\link{get_hfreferral_form_query}()},
\code{\link{get_historical_hf_diag_query}()},
\code{\link{get_previous_encounter_query}()},
\code{\link{get_referred_to_facility_query}()}
}
\concept{query}
