% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmanagement.R
\name{get_precipitants}
\alias{get_precipitants}
\title{Processes Precipitant information in enrolment form}
\usage{
get_precipitants(df)
}
\arguments{
\item{df}{a dataframe containing enrolment forms}
}
\description{
Converts the precipiant fields in the enrolment into separate categories
}
\seealso{
\code{\link[=process_hfmanagement_form]{process_hfmanagement_form()}} for df,
\code{\link[=get_aetiology]{get_aetiology()}} for processing aetiology field

Other hfmanagement: 
\code{\link{get_aetiology}()},
\code{\link{get_ecg_data}()},
\code{\link{get_ecg_on_admit}()},
\code{\link{process_hfmanagement_form}()}
}
\concept{hfmanagement}
