% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_previous_encounter_query}
\alias{get_previous_encounter_query}
\title{rank all inpatient encounters by date}
\usage{
get_previous_encounter_query(
  df,
  PERSON_ID = "PERSON_ID",
  ENCNTR_ID = "ENCNTR_ID"
)
}
\description{
rank all inpatient encounters by date
}
