% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{journey_length}
\alias{journey_length}
\title{Set start/end times for journeys}
\usage{
journey_length(
  journeys,
  ...,
  .default_dots = rlang::exprs(journey_key = JOURNEY_KEY, journey_start =
    JOURNEY_START, journey_end = JOURNEY_END, journey_days = JOURNEY_DAYS, admit_dttm =
    ADMIT_DTTM, discharge_dttm = DISCH_DTTM, )
)
}
\arguments{
\item{journeys}{the journey df}

\item{...}{see section Column Specification}
}
\value{
the df, but w/ start, end, and LOS in days
}
\description{
Set start/end times for journeys
}
\seealso{
Other journey_building: 
\code{\link{create_journey_table}()},
\code{\link{journey_order}()},
\code{\link{link_encounters_by_timestamps}()},
\code{\link{link_encs_in_groups}()},
\code{\link{relink_apac_to_journeys}()},
\code{\link{terminate_journeys_by}()}
}
\concept{journey_building}
\keyword{internal}
