% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{journey_order}
\alias{journey_order}
\title{Give a order for each encounter based on their journey key in ascending order
e.g. for a journey with 3 encounters, the first encounter will have rank 1,
second with rank 2 and so on}
\usage{
journey_order(
  journeys,
  ...,
  .default_dots = rlang::exprs(journey_key = JOURNEY_KEY, discharge_dttm = DISCH_DTTM,
    encntr_key = ENCNTR_ID, updt_dttm = UPDT_DT_TM)
)
}
\arguments{
\item{journeys}{the journey df}

\item{...}{see section Column Specification}
}
\value{
the df, journey_order, the final encounter
}
\description{
Give a order for each encounter based on their journey key in ascending order
e.g. for a journey with 3 encounters, the first encounter will have rank 1,
second with rank 2 and so on
}
\seealso{
Other journey_building: 
\code{\link{create_journey_table}()},
\code{\link{journey_length}()},
\code{\link{link_encounters_by_timestamps}()},
\code{\link{link_encs_in_groups}()},
\code{\link{relink_apac_to_journeys}()},
\code{\link{terminate_journeys_by}()}
}
\concept{journey_building}
\keyword{internal}
