% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{link_encounters_by_timestamps}
\alias{link_encounters_by_timestamps}
\alias{link_encounters_to_journeys}
\title{Link Encounters Into Journeys}
\usage{
link_encounters_by_timestamps(
  df,
  ...,
  max_merge_gap_hours = 24,
  .default_dots = rlang::exprs(person_key = person_key, encntr_key = encntr_key,
    journey_key = JOURNEY_KEY)
)
}
\arguments{
\item{df}{The dataframe with rows carrying PID, EID, admit/disch datetimes
Should be cleaned by cleaner first (add example)}

\item{...}{see section Column Specifications}

\item{max_merge_gap_hours}{The maximum number of hours between
the end of one encounter, and the start of the next, such that they still
be grouped as one journey}
}
\value{
A dataframe with an additional variable, JOURNEY_KEY which groups
corresponds to the ENCNTR_KEY of the earliest encounter within the journey.
additionaly, has the following cols: \code{JOURNEY_START}, \code{JOURNEY_END},
\code{JOURNEY_DAYS}
}
\description{
Accepts df of encounters, and groups them by JOURNEY_KEY.
This JOURNEY_KEY groups sets of encounters according to whether they
overlap/their proximity in time
}
\seealso{
Other journey_building: 
\code{\link{create_journey_table}()},
\code{\link{journey_length}()},
\code{\link{journey_order}()},
\code{\link{link_encs_in_groups}()},
\code{\link{relink_apac_to_journeys}()},
\code{\link{terminate_journeys_by}()}
}
\concept{journey_building}
