% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{link_encs_in_groups}
\alias{link_encs_in_groups}
\title{Link encounters inside a group}
\usage{
link_encs_in_groups(
  encs,
  ...,
  max_merge_gap_hours = 24,
  .default_dots = rlang::exprs(journey_key = JOURNEY_KEY, encntr_key = ENCNTR_ID,
    admit_dttm = BEG_EFFECTIVE_DT_TM, discharge_dttm = DISCH_DT_TM)
)
}
\description{
Link encounters inside a group
}
\seealso{
Other journey_building: 
\code{\link{create_journey_table}()},
\code{\link{journey_length}()},
\code{\link{journey_order}()},
\code{\link{link_encounters_by_timestamps}()},
\code{\link{relink_apac_to_journeys}()},
\code{\link{terminate_journeys_by}()}
}
\concept{journey_building}
\keyword{internal}
