% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_db}
\alias{load_db}
\title{Load a table from database}
\usage{
load_db(table_name)
}
\arguments{
\item{table_name}{a string, name of the tabke}
}
\description{
Load a table from database
}
\seealso{
Other data-utility: 
\code{\link{execute_query}()},
\code{\link{execute_write_to_db}()},
\code{\link{get_db_connection}()},
\code{\link{get_first_non_na}()},
\code{\link{load_data}()},
\code{\link{str_to_datetime}()}
}
\concept{data-utility}
