% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_of_sep.R
\name{mode_of_separation_rules}
\alias{mode_of_separation_rules}
\alias{mode_of_sep_groupings}
\alias{enc_sep_base_patterns}
\alias{enc_sep_transfer_patterns}
\title{Rules for mode of separation}
\usage{
mode_of_sep_groupings()

enc_sep_base_patterns(
  first_pass = mode_sep_first,
  transfer = mode_sep_transfer,
  death = mode_sep_death,
  dests = destination,
  simple_vals = c("death", "dama", "home"),
  transfer_vals = c("transfer")
)

enc_sep_transfer_patterns(transfer = mode_sep_transfer, dests = destination)
}
\description{
Rules for mode of separation
}
\section{Base Rules}{

The following patterns/rules are used to relate the extracted top level
classifications to an output mode of separation.
}

\section{Transfer Rules}{

The following rules are called from the enc_sep_case_transfer function,
as part of the evaluation of the base ruleset.
}

\seealso{
Other mode_of_separation: 
\code{\link{encounter_classify_separation}()}
}
\concept{encounter}
\concept{mode_of_separation}
