% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfenrolment.R
\name{process_hfenrolment_form}
\alias{process_hfenrolment_form}
\title{Pivots the HF enrolment form}
\usage{
process_hfenrolment_form()
}
\description{
Pivots the HF enrolment forms from long form to wide
based on the fields given in \code{get_enrolment_fields()}
for each encounter.
For encounters with multiple enrolment forms,
the most recent, non-null result will be used
}
\seealso{
\code{\link[=get_enrolment_fields]{get_enrolment_fields()}} for list of fields extracts,
\code{\link[=get_first_non_na]{get_first_non_na()}} for definition for non empty fields, and
\code{\link[=get_hfenrolment_form_query]{get_hfenrolment_form_query()}} for the dataframe that will be pivoted

Other hfenrolment: 
\code{\link{get_enrolment_fields}()},
\code{\link{get_n_precipitants}()},
\code{\link{process_hfmanagement}()},
\code{\link{summarise_enrolment}()}
}
\concept{hfenrolment}
