% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfenrolment.R
\name{process_hfmanagement}
\alias{process_hfmanagement}
\title{Process HF enrolment forms}
\usage{
process_hfmanagement(df)
}
\arguments{
\item{df}{the dataframe to processed}
}
\description{
Processes the fields of the HF enrolment forms
Fields list \code{RESULT_VAL_ECGONADMISSION},
\code{RESULT_VAL_HFPRECIPITAMTS}, \code{RESULT_VAL_AETIOLOGYOFCCF}
will be seperated into different columns based on their value
LVEF will be binned accordingly
The number of precipiants will be calculated
}
\seealso{
\code{\link[=process_hfenrolment_form]{process_hfenrolment_form()}} for more details on the dataframe used

Other hfenrolment: 
\code{\link{get_enrolment_fields}()},
\code{\link{get_n_precipitants}()},
\code{\link{process_hfenrolment_form}()},
\code{\link{summarise_enrolment}()}
}
\concept{hfenrolment}
