% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmanagement.R
\name{process_hfmanagement_form}
\alias{process_hfmanagement_form}
\title{Processes HF enrolment forms}
\usage{
process_hfmanagement_form(forms, dcp_forms_activity)
}
\arguments{
\item{df}{a dataframe containing enrolment forms}
}
\description{
Processes HF enrolment forms
}
\seealso{
Other hfmanagement: 
\code{\link{get_aetiology}()},
\code{\link{get_ecg_data}()},
\code{\link{get_ecg_on_admit}()},
\code{\link{get_precipitants}()}
}
\concept{hfmanagement}
