% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfreferral.R
\name{process_hfreferral_form}
\alias{process_hfreferral_form}
\title{Create Referral form df}
\usage{
process_hfreferral_form()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\concept{hfreferral}
