% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{process_homevisit_form}
\alias{process_homevisit_form}
\title{Note these grouping are done based on `PARENT_EVENT_ID`/`PARENT_ENTITY_ID`
and NOT `ENCNTR_ID`. Al these visit forms are linked to the same encounter}
\usage{
process_homevisit_form(forms = NULL, dcp_forms_activity = NULL)
}
\description{
Note these grouping are done based on `PARENT_EVENT_ID`/`PARENT_ENTITY_ID`
and NOT `ENCNTR_ID`. Al these visit forms are linked to the same encounter
}
