% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{process_inpatient}
\alias{process_inpatient}
\title{Creates the dataframe for encounters}
\usage{
process_inpatient(df)
}
\description{
Creates the dataframe for encounters
}
\seealso{
\code{\link[=process_inpatient]{process_inpatient()}} for df

Other hfinpatient: 
\code{\link{extract_age_los}()},
\code{\link{extract_alert}()},
\code{\link{extract_diagnosis}()},
\code{\link{extract_discharge_meds}()},
\code{\link{extract_echos}()},
\code{\link{extract_form_result}()},
\code{\link{extract_hfenrolment_form}()},
\code{\link{extract_hfreferral_form}()},
\code{\link{extract_path_result}()},
\code{\link{extract_patho_result}()},
\code{\link{get_most_recent_order}()}
}
\concept{hfinpatient}
