% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default_dots}
\alias{set_default_dots}
\title{Assign defaults to dotted inputs}
\usage{
set_default_dots(.default_dots, ...)
}
\arguments{
\item{.default_dots}{a named list of (typically) expressions, that specify
default values for args in `...` if they aren't given by `...`}
}
\value{
the input `...`,  with any named args from `.default_dots`
        missing from `...` being set by `defaults`.
}
\description{
`...` is a list of arguments, usually coming from `...` passed directly to
 the function.
}
\keyword{internal}
