% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{split_on_finals}
\alias{split_on_finals}
\title{Split journeys on FINAL_ENC == TRUE}
\usage{
split_on_finals(group)
}
\arguments{
\item{group}{the group given by a dplyr::group_by %>% dplyr::do
pipe on which to operate}
}
\description{
Given a df w/ JOURNEY_KEY, ENCNTR_KEY, FINAL_ENC cols
   Where FINAL ENC is a logical, terminate jouneys and
   begin new ones when you see FINAL_ENC == true
}
\keyword{internal}
