% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_datetime}
\alias{str_to_datetime}
\title{Convert strings column to datetime format}
\usage{
str_to_datetime(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Converts any column containing _dttm or date_time
into datetime format
}
\seealso{
Other data-utility: 
\code{\link{execute_query}()},
\code{\link{execute_write_to_db}()},
\code{\link{get_db_connection}()},
\code{\link{get_first_non_na}()},
\code{\link{load_data}()},
\code{\link{load_db}()}
}
\concept{data-utility}
