% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfenrolment.R
\name{summarise_enrolment}
\alias{summarise_enrolment}
\title{Get most recent and first enrolment}
\usage{
summarise_enrolment(df)
}
\description{
Gets the most and first enrolment for each person
}
\seealso{
\code{\link[=process_hfenrolment_form]{process_hfenrolment_form()}} for more details
on the df used

Other hfenrolment: 
\code{\link{get_enrolment_fields}()},
\code{\link{get_n_precipitants}()},
\code{\link{process_hfenrolment_form}()},
\code{\link{process_hfmanagement}()}
}
\concept{hfenrolment}
