% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{terminate_journeys_by}
\alias{terminate_journeys_by}
\title{Terminate Journeys on encounters having particular values}
\usage{
terminate_journeys_by(journeys, ...)
}
\arguments{
\item{journeys}{the dataframe representing the journey-keyed encounters}

\item{...}{Conditions to pass on to check an encounter should terminate
its journey}
}
\value{
The df, but with the journey key reworked to reflect the
termination. Journey length, etc, should also be affected.
}
\description{
Helper function for second pass at journey building.
This time, validates / Checks that encounters having characteristics
that indicate the encounter should terminate its journey
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\section{TODO}{

Set this as internal, wrap it in something
}

\examples{
\dontrun{
journeys <- encounters \%>\%
   link_encounters_by_timestamps()

dispositions <- c("Departed - Treatment Completed", "Death Type 1")
rekeyed <- journeys \%>\%
   terminate_journeys_by(DISCH_DISPOSITION \%in\% dispositions)
}
}
\seealso{
Other journey_building: 
\code{\link{create_journey_table}()},
\code{\link{journey_length}()},
\code{\link{journey_order}()},
\code{\link{link_encounters_by_timestamps}()},
\code{\link{link_encs_in_groups}()},
\code{\link{relink_apac_to_journeys}()}

Other deprecated: 
\code{\link{extract_hf_hist_diag}()},
\code{\link{get_previous_encounter_query}()}
}
\concept{deprecated}
\concept{journey_building}
