% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_db}
\alias{write_db}
\title{Write a dataframe to a database}
\usage{
write_db(df, con, table_name)
}
\arguments{
\item{df}{a dataframe of the data to write}

\item{con, }{the connection to the database}

\item{a}{string, name of the table to write to}
}
\description{
Write a dataframe to a database
}
