% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_of_sep.R
\name{encounter_udf_add_dest}
\alias{encounter_udf_add_dest}
\title{classifies the destination given by
\verb{Referred to Facility}}
\usage{
encounter_udf_add_dest(df, value_col = "VALUE_CD")
}
\description{
classifies the destination given by
\verb{Referred to Facility} value in \code{ENCOUNTER_UDF} table
to either private,racf or the given public hospital
}
