% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_alert}
\alias{extract_alert}
\title{extracts advanced care directives and checks whether the alert was
created before or during the encounter}
\usage{
extract_alert(df)
}
\description{
extracts advanced care directives and checks whether the alert was
created before or during the encounter
}
