% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_discharge_meds}
\alias{extract_discharge_meds}
\title{create discharge medications from encounters}
\usage{
extract_discharge_meds(df)
}
\description{
create discharge medications from encounters
}
