% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_form_result}
\alias{extract_form_result}
\title{extracts the result and bin pathology result}
\usage{
extract_form_result(
  df,
  regex,
  prefix,
  breaks = NULL,
  labels = NULL,
  is_numeric = FALSE
)
}
\description{
extracts the result and bin pathology result
}
