% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{extract_hf_hist_diag}
\alias{extract_hf_hist_diag}
\title{extract the latest hf related icd10 diagnosis for each encounter}
\usage{
extract_hf_hist_diag(df)
}
\description{
extract the latest hf related icd10 diagnosis for each encounter
}
