% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_acd_problem_query}
\alias{get_acd_problem_query}
\title{queries the first instance
the advance care directive for a given `PERSON_ID`
extracts the earliest record of the directive}
\usage{
get_acd_problem_query()
}
\description{
queries the first instance
the advance care directive for a given `PERSON_ID`
extracts the earliest record of the directive
}
