% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_discharge_referral}
\alias{get_discharge_referral}
\title{get the most recently updated note for each encounter}
\usage{
get_discharge_referral()
}
\description{
get the most recently updated note for each encounter
}
