% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_hf_alert_query}
\alias{get_hf_alert_query}
\title{queries the first instance
the advance care directive for a given `PERSON_ID`}
\usage{
get_hf_alert_query()
}
\description{
queries the first instance
the advance care directive for a given `PERSON_ID`
}
