% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfinpatient.R
\name{get_most_recent_order}
\alias{get_most_recent_order}
\title{extract the most recent event/order based on the given time column}
\usage{
get_most_recent_order(
  df,
  order_df,
  suffix,
  order_dttm_col,
  df_dttm = "DISCH_DT_TM",
  delta_cutoff = 0
)
}
\description{
a helper function to extract the most recent event/order
based on the given time column
}
