% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfhomevisit.R
\name{process_homevisit_form}
\alias{process_homevisit_form}
\title{Process homevisit forms}
\usage{
process_homevisit_form(forms = NULL, dcp_forms_activity = NULL)
}
\description{
Note these grouping are done based on \code{PARENT_EVENT_ID}/\code{PARENT_ENTITY_ID}
and NOT \code{ENCNTR_ID}. Al these visit forms are linked to the same encounter
}
