% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default_dots}
\alias{set_default_dots}
\title{Assign defaults to dotted inputs}
\usage{
set_default_dots(.default_dots, ...)
}
\arguments{
\item{.default_dots}{a named list of (typically) expressions, that specify
default values for args in \code{...} if they aren't given by \code{...}}
}
\value{
the input \code{...},  with any named args from \code{.default_dots}
missing from \code{...} being set by \code{defaults}.
}
\description{
\code{...} is a list of arguments, usually coming from \code{...} passed directly to
the function.
}
\keyword{internal}
